Trekk felles testutils ut i egen fil
