+++ src/test/kotlin/no/nav/tilleggsstonader/soknad/soknad/SøknadTestUtil.kt
@@ -1,8 +1,17 @@
 package no.nav.tilleggsstonader.soknad.soknad
 
+import no.nav.tilleggsstonader.kontrakter.felles.Hovedytelse
 import no.nav.tilleggsstonader.kontrakter.felles.ObjectMapperProvider.objectMapper
 import no.nav.tilleggsstonader.kontrakter.felles.Stønadstype
+import no.nav.tilleggsstonader.kontrakter.søknad.DatoFelt
+import no.nav.tilleggsstonader.kontrakter.søknad.EnumFelt
+import no.nav.tilleggsstonader.kontrakter.søknad.EnumFlereValgFelt
+import no.nav.tilleggsstonader.kontrakter.søknad.JaNei
+import no.nav.tilleggsstonader.kontrakter.søknad.SelectFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.Søknadsskjema
+import no.nav.tilleggsstonader.kontrakter.søknad.VerdiFelt
+import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.TypePengestøtte
+import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.ÅrsakOppholdUtenforNorge
 import no.nav.tilleggsstonader.libs.utils.osloNow
 import no.nav.tilleggsstonader.soknad.infrastruktur.database.JsonWrapper
 import no.nav.tilleggsstonader.soknad.person.dto.Barn
@@ -42,4 +51,51 @@ object SøknadTestUtil {
             )
         }
     }
+
+    private fun arbeidOgOppholdDto() = ArbeidOgOppholdDto(
+        jobberIAnnetLand = EnumFelt("Jobber du i et annet land enn Norge?", JaNei.JA, "Ja", emptyList()),
+        jobbAnnetLand = SelectFelt("Hvilket land jobber du i?", "SWE", "Sverige"),
+        harPengestøtteAnnetLand = EnumFlereValgFelt(
+            "Mottar du pengestøtte fra et annet land enn Norge?",
+            listOf(VerdiFelt(TypePengestøtte.SYKEPENGER, "Sykepenger")),
+            listOf("Sykepenger", "Annet"),
+        ),
+        pengestøtteAnnetLand = SelectFelt("Hvilket land mottar du pengestøtte fra?", "SWE", "Sverige"),
+        harOppholdUtenforNorgeSiste12mnd = EnumFelt(
+            "Jobber du i et annet land enn Norge?",
+            JaNei.JA,
+            "Ja",
+            emptyList(),
+        ),
+        oppholdUtenforNorgeSiste12mnd = listOf(oppholdUtenforNorgeDto(), oppholdUtenforNorgeDto()),
+        harOppholdUtenforNorgeNeste12mnd = EnumFelt(
+            "Jobber du i et annet land enn Norge?",
+            JaNei.JA,
+            "Ja",
+            emptyList(),
+        ),
+        oppholdUtenforNorgeNeste12mnd = listOf(oppholdUtenforNorgeDto()),
+    )
+
+    fun hovedytelseDto() = HovedytelseDto(
+        ytelse = EnumFlereValgFelt(
+            label = "Hovedytelse?",
+            verdier = listOf(
+                VerdiFelt(Hovedytelse.AAP, "AAP"),
+                VerdiFelt(Hovedytelse.OVERGANGSSTØNAD, "Overgangsstønad"),
+            ),
+            alternativer = listOf("Alt1", "Alt2"),
+        ),
+        arbeidOgOpphold = arbeidOgOppholdDto(),
+    )
+    private fun oppholdUtenforNorgeDto() = OppholdUtenforNorgeDto(
+        land = SelectFelt("Hvilket land har du oppholdt deg i?", "SWE", "Sverige"),
+        årsak = EnumFlereValgFelt(
+            "Hva gjorde du i dette landet?",
+            listOf(VerdiFelt(ÅrsakOppholdUtenforNorge.JOBB, "Jobb")),
+            alternativer = listOf("Jobb", "Studier"),
+        ),
+        fom = DatoFelt("Fom", LocalDate.of(2024, 1, 1)),
+        tom = DatoFelt("Fom", LocalDate.of(2024, 1, 1)),
+    )
 }
