+++ src/test/kotlin/no/nav/tilleggsstonader/soknad/soknad/barnetilsyn/SøknadBarnetilsynUtil.kt
@@ -1,36 +1,21 @@
 package no.nav.tilleggsstonader.soknad.soknad.barnetilsyn
 
-import no.nav.tilleggsstonader.kontrakter.felles.Hovedytelse
-import no.nav.tilleggsstonader.kontrakter.søknad.DatoFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.Dokument
 import no.nav.tilleggsstonader.kontrakter.søknad.DokumentasjonFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.EnumFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.EnumFlereValgFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.JaNei
-import no.nav.tilleggsstonader.kontrakter.søknad.SelectFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.Vedleggstype
 import no.nav.tilleggsstonader.kontrakter.søknad.VerdiFelt
 import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.AnnenAktivitetType
 import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.TypeBarnepass
-import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.TypePengestøtte
 import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.ÅrsakBarnepass
-import no.nav.tilleggsstonader.kontrakter.søknad.barnetilsyn.ÅrsakOppholdUtenforNorge
-import java.time.LocalDate
+import no.nav.tilleggsstonader.soknad.soknad.SøknadTestUtil
 import java.util.UUID
 
 object SøknadBarnetilsynUtil {
     val søknad = SøknadBarnetilsynDto(
-        hovedytelse = HovedytelseDto(
-            ytelse = EnumFlereValgFelt(
-                label = "Hovedytelse?",
-                verdier = listOf(
-                    VerdiFelt(Hovedytelse.AAP, "AAP"),
-                    VerdiFelt(Hovedytelse.OVERGANGSSTØNAD, "Overgangsstønad"),
-                ),
-                alternativer = listOf("Alt1", "Alt2"),
-            ),
-            arbeidOgOpphold = arbeidOgOppholdDto(),
-        ),
+        hovedytelse = SøknadTestUtil.hovedytelseDto(),
         aktivitet = Aktivitet(
             aktiviteter = EnumFlereValgFelt(
                 "Hvilken aktivitet søker du om støtte i forbindelse med?",
@@ -56,42 +41,6 @@ object SøknadBarnetilsynUtil {
             ),
         ),
         dokumentasjon = listOf(lagDokumentasjonFelt(), lagDokumentasjonFeltBarn()),
-    )
-
-    private fun arbeidOgOppholdDto() = ArbeidOgOppholdDto(
-        jobberIAnnetLand = EnumFelt("Jobber du i et annet land enn Norge?", JaNei.JA, "Ja", emptyList()),
-        jobbAnnetLand = SelectFelt("Hvilket land jobber du i?", "SWE", "Sverige"),
-        harPengestøtteAnnetLand = EnumFlereValgFelt(
-            "Mottar du pengestøtte fra et annet land enn Norge?",
-            listOf(VerdiFelt(TypePengestøtte.SYKEPENGER, "Sykepenger")),
-            listOf("Sykepenger", "Annet"),
-        ),
-        pengestøtteAnnetLand = SelectFelt("Hvilket land mottar du pengestøtte fra?", "SWE", "Sverige"),
-        harOppholdUtenforNorgeSiste12mnd = EnumFelt(
-            "Jobber du i et annet land enn Norge?",
-            JaNei.JA,
-            "Ja",
-            emptyList(),
-        ),
-        oppholdUtenforNorgeSiste12mnd = listOf(oppholdUtenforNorgeDto(), oppholdUtenforNorgeDto()),
-        harOppholdUtenforNorgeNeste12mnd = EnumFelt(
-            "Jobber du i et annet land enn Norge?",
-            JaNei.JA,
-            "Ja",
-            emptyList(),
-        ),
-        oppholdUtenforNorgeNeste12mnd = listOf(oppholdUtenforNorgeDto()),
-    )
-
-    private fun oppholdUtenforNorgeDto() = OppholdUtenforNorgeDto(
-        land = SelectFelt("Hvilket land har du oppholdt deg i?", "SWE", "Sverige"),
-        årsak = EnumFlereValgFelt(
-            "Hva gjorde du i dette landet?",
-            listOf(VerdiFelt(ÅrsakOppholdUtenforNorge.JOBB, "Jobb")),
-            alternativer = listOf("Jobb", "Studier"),
-        ),
-        fom = DatoFelt("Fom", LocalDate.of(2024, 1, 1)),
-        tom = DatoFelt("Fom", LocalDate.of(2024, 1, 1)),
     )
 
     private fun lagDokumentasjonFelt() = DokumentasjonFelt(
